# POC of using easyjson for serializing entities