… server/go.sum
… server/public/model/post_easyjson.go
… server/public/model/websocket_message_easyjson.go