#### Summary

This is a proof of concept to use easyjson for serializing entities. I only added it for websocket events and for posts, but we could do this same thing for other entities, not necesarilly all of them, we can focus only on the most common ones.

I added a serialization bencharmark for post and reused the existing bencharmks from websocket events, here is the comparison:

```
goos: linux
goarch: amd64
pkg: github.com/mattermost/mattermost/server/public/model
cpu: Intel(R) Core(TM) i9-10900X CPU @ 3.70GHz
                                                            │ benchmark_without_easyjson.txt │     benchmark_with_easyjson.txt     │
                                                            │             sec/op             │    sec/op     vs base               │
PostToJSON-20                                                                   7.831µ ± ∞ ¹   3.147µ ± ∞ ¹  -59.81% (p=0.008 n=5)
WebSocketEvent_ToJSON/SerializedNTimes-20                                      113.71µ ± ∞ ¹   50.78µ ± ∞ ¹  -55.34% (p=0.008 n=5)
WebSocketEvent_ToJSON/PrecomputedNTimes-20                                      84.58µ ± ∞ ¹   78.50µ ± ∞ ¹        ~ (p=0.310 n=5)
WebSocketEvent_ToJSON/PrecomputedAndSerializedNTimes-20                         206.8µ ± ∞ ¹   142.8µ ± ∞ ¹  -30.96% (p=0.008 n=5)
WebSocketEvent_ToJSON/PrecomputedOnceAndSerializedNTimes-20                    118.38µ ± ∞ ¹   57.58µ ± ∞ ¹  -51.36% (p=0.008 n=5)
geomean                                                                         71.31µ         40.06µ        -43.83%
¹ need >= 6 samples for confidence interval at level 0.95
```