@jespino 2024-09-20
@agnivade, an option to reduce CPU consumption is to save a bit on serialization, which I guess is something that we do a lot. this is a POC, and the local benchmark gives good results.