@agnivade 2024-09-23
Thanks @jespino - I have the CPU profiles from a 100K test. Currently, there's not a particular area to improve specifically. And JSON marshalling doesn't appear that prominently in the profile. 

Two major areas where time is spent are:
webConn.ShouldSendEvent: 15.2% 
webConn.WritePump: 17.1% 

Attaching the profile below.

[cpu_100K_redis (copy).zip](https://github.com/user-attachments/files/17093807/cpu_100K_redis.copy.zip)

However, I think we can definitely include this atleast for websocket events. cc @streamer45 @agarciamontoro for some thoughts.