@agnivade 2024-09-23
Yep, I think we can definitely include this for websocket events.