@streamer45 2024-09-23
First of all, thanks for looking into this. I think this exploration is a great starting point.

That said, I feel that parting ways from such a core (standard library) feature should not be taken lightly and I'd expect this work to go through a more formal proposal process that only culminates with a PR. 

Data serialization has important aspects to consider that go well beyond raw performance (e.g., security, maintainability) so I'd expect us to be thorough and explore all these angles.

I'd also expect us to consider more alternatives, especially those referenced by [JSON v2 benchmarks](https://github.com/go-json-experiment/jsonbench?tab=readme-ov-file#marshal-performance).

I propose we start a discussion internally and plan some next steps.