@jespino 2024-09-23
Thanks @streamer45! Yes I agree that this shouldn't be decided through this pr, this was a medium to share the idea, but I didn't expect to be the final solution. I only want to share the fact that there are some space for improvement on this specific problem and the overall performance implications of reducing the number of allocations and the pressure from the garbage collector can be interesting. The final solution can be easyjson or can be any other that we encounter useful.