@agnivade 2024-09-24
Yeah, worth highlighting that the Go team is working on a json/v2 implementation as Claudio pointed out. We should definitely do a thorough test of running benchmarks for all the serializers in market and come to a conclusion.